feat: secu

+++ LegoList.php
@@ -11,16 +11,16 @@
     </head>
     <body>
         <header>
+          <?php session_start();?>
         <nav class="topnav" id="myTopnav">
                 <ul>
                   <li class="title">Motify</li>
-                  <li><a href="index.html">Home</a></li>
-                  <li><a href="newUser.php">New user</a></li>
+                  <li><a href="index.php">Home</a></li>
+                  <?php if((isset($_SESSION['Role']))&&($_SESSION['Role'] == 'ADMIN')){echo'<li><a href="newUser.php">New user</a></li>';}?>
                   <li><a href="UserList.php">Liste des Utilisateurs</a></li>
-                  <li><a href="newLego.php">New lego</a></li>
+                  <?php if((isset($_SESSION['Role']))&&($_SESSION['Role'] == 'ADMIN')){echo'<li><a href="newLego.php">New lego</a></li>';}?>
                   <li><a href="LegoList.php">Liste des lego</a></li>
                   <?php
-                    session_start();
                     if (!$_SESSION['connecter'] == TRUE) {
                         header('Location: connection.php');
                         }else{
@@ -59,8 +59,10 @@
             echo "<td>".$article->getBoite()."</td>";
             echo "<td>".$article->getNotice()."</td>";
             echo "<td><a href='afficher.php?id=".$article->getLego_id()."'>afficher</a></td>";
+            if((isset($_SESSION['Role']))&&($_SESSION['Role'] == 'ADMIN')){
             echo "<td><a href='updateLego.php?id=".$article->getLego_id()."'>Modifier</a></td>";
             echo "<td><a href='deleteLego.php?id=".$article->getLego_id()."'>Suprimer</a></td>";
+            }
             echo "</tr>";
             }
             ?>
